cesem/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap3==2022.1
dj-database-url==2.0.0
Django==4.2
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-datatables==0.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
pep8==1.7.1
platformdirs==3.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
sentry-sdk==1.32.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
xlrd==2.0.1
